+++ user-profile-boundary/UserProfile.tsx
@@ -1,49 +1,17 @@
 import { AvatarView } from "@components/Avatar"
 import { Caption, Headline, Subtitle } from "@components/Text"
-import { EventCard } from "@event/EventCard"
-import { featureContext } from "@lib/FeatureContext"
 import { useQuery, useQueryClient } from "@tanstack/react-query"
-import { AlphaUserStorage } from "@user/alpha"
 import { FriendRequestButton, useFriendRequest } from "@user/FriendRequest"
 import React from "react"
 import { StyleProp, StyleSheet, View, ViewStyle } from "react-native"
 import { FlatList } from "react-native-gesture-handler"
-import { TiFAPI } from "TiFShared/api"
 import {
   UnblockedUserRelationsStatus,
-  UserID,
-  UserProfile
+  UserID
 } from "TiFShared/domain-models/User"
+import { UserProfileDisplayInfo, UserProfileFeature } from "./Context"
 import { userProfileQueryKey } from "./QueryKey"
-import { userUpcomingEvents, useUpcomingEvents } from "./UpcomingEvents"
-
-export type BasicUser = Omit<UserProfile, "createdDateTime" | "updatedDateTime">
-
-export type UserProfileResult =
-  | {
-      status: "success"
-      user: BasicUser
-    }
-  | { status: "blocked-you" | "user-not-found" }
-
-export const userProfile = async (
-  userId: UserID,
-  api: TiFAPI = TiFAPI.productionInstance,
-  storage: AlphaUserStorage = AlphaUserStorage.default
-): Promise<UserProfileResult> => {
-  const storageData = await storage.user()
-  if (storageData && storageData.id === userId) {
-    return {
-      status: "success",
-      user: { ...storageData, relationStatus: "current-user" }
-    }
-  }
-  const resp = await api.getUser({ params: { userId } })
-  if (resp.status === 200) {
-    return { status: "success", user: resp.data as BasicUser }
-  }
-  return { status: resp.data.error }
-}
+import { useUpcomingEvents } from "./UpcomingEvents"
 
 export type UseUserProfileEnvironment = {
   userId: UserID
@@ -70,11 +38,6 @@ export const useUserProfile = ({ userId }: UseUserProfileEnvironment) => {
     return { status: query.status }
   }
 }
-
-export const UserProfileFeature = featureContext({
-  fetchUserProfile: userProfile,
-  fetchUpcomingEvents: userUpcomingEvents
-})
 
 export type UserProfileProps = {
   userInfoState: ReturnType<typeof useUserProfile>
@@ -103,7 +66,7 @@ export const UserProfileView = ({
             ? upcomingEventsState.data.events
             : undefined
         }
-        renderItem={({ item }) => <EventCard event={item} />}
+        renderItem={({ item }) => <Headline>{"Idk"}</Headline>}
         ListHeaderComponent={
           <View>
             <UserInfoView
@@ -149,7 +112,7 @@ export const UserInfoView = ({
 }
 
 export type BaseUserInfoViewProps = {
-  user: BasicUser
+  user: UserProfileDisplayInfo
   onRelationStatusChanged: (
     id: UserID,
     status: UnblockedUserRelationsStatus
@@ -157,7 +120,7 @@ export type BaseUserInfoViewProps = {
   style?: StyleProp<ViewStyle>
 }
 
-export const BaseUserInfoView = ({
+const BaseUserInfoView = ({
   user,
   onRelationStatusChanged,
   style
